requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
flask==2.3.2
flask-cors==4.0.0
tqdm==4.65.0